% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asymptotic_stats.R, R/data.R
\docType{data}
\name{asymptotic_stats}
\alias{asymptotic_stats}
\title{Asymptotic stats}
\format{
A vector with elements ...

A numeric vector of statistics
}
\source{
Generated from simulation study
}
\usage{
data(asymptotic_stats)

asymptotic_stats
}
\description{
50,000 simulated values from the asymptotic distribution of the single change point statistic

Dataset containing pre-computed statistics for asymptotic testing in change point detection.
}
\keyword{datasets}
