% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KDist.R
\name{bw_optim}
\alias{bw_optim}
\title{Calculate optimal bandwidth}
\usage{
bw_optim(x, sample_limit = 1000, scale_factor = 0.5, group = NULL)
}
\arguments{
\item{x}{Matrix or vector of data}

\item{sample_limit}{Maximum number of samples to use for calculation}

\item{scale_factor}{Scaling factor for the bandwidth}

\item{group}{Optional grouping of variables}
}
\value{
Bandwidth parameter value or vector
}
\description{
Calculate optimal bandwidth
}
\examples{
# Working with a data frame (must have all numeric columns)
df <- data.frame(a = rnorm(20), b = rnorm(20), c = rnorm(20))
bw_df <- bw_optim(df)

# The function will throw an error if non-numeric columns are present
\dontrun{
df_with_factors <- data.frame(a = rnorm(20), b = as.factor(1:20))
bw_error <- bw_optim(df_with_factors) # This will error
}
}
