% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmanova.R
\name{dmanova2}
\alias{dmanova2}
\title{Kernel and Distance-Based MANOVA}
\usage{
dmanova2(
  formula,
  data = NULL,
  type_rep = "euclidean",
  type_cov = "euclidean",
  expo_rep = 1,
  expo_cov = 1,
  n_perm = 299,
  parallel = FALSE,
  n_cores = NULL,
  contr.unordered = "contr.sum",
  contr.ordered = "contr.poly"
)
}
\arguments{
\item{formula}{A formula with the response on the left and predictors on the right}

\item{data}{An optional data frame containing the variables in the formula}

\item{type_rep}{Type of distance for response}

\item{type_cov}{Type of distance for covariates}

\item{expo_rep}{Exponent parameter for response}

\item{expo_cov}{Exponent parameter for covariates}

\item{n_perm}{Number of permutations}

\item{parallel}{Logical; whether to use parallel processing}

\item{n_cores}{Number of cores for parallel processing}

\item{contr.unordered}{Contrast method for unordered factors}

\item{contr.ordered}{Contrast method for ordered factors}
}
\value{
A list containing test statistics and p-value
}
\description{
An alternative implementation of kernel and distance-based multivariate analysis of variance.
All the variables will be transformed into kernel or distance matrices.
}
