% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_point.R
\name{.generate_random_intervals}
\alias{.generate_random_intervals}
\title{Generate random intervals and compute test statistics}
\usage{
.generate_random_intervals(full_dist, type, start, end, M, num_cores = NULL)
}
\arguments{
\item{full_dist}{Full distance matrix}

\item{type}{Type of distance/kernel}

\item{start}{Start index of overall interval}

\item{end}{End index of overall interval}

\item{M}{Number of random intervals to generate}

\item{num_cores}{Number of cores for parallel processing}
}
\value{
Matrix with 4 rows and M columns (start, end, change point, test statistic)
}
\description{
Generate random intervals and compute test statistics
}
\keyword{internal}
