% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_point.R
\name{.remove_duplicates}
\alias{.remove_duplicates}
\title{Remove duplicate columns from a matrix}
\usage{
.remove_duplicates(M)
}
\arguments{
\item{M}{Input matrix}
}
\value{
Matrix with duplicate columns removed
}
\description{
Remove duplicate columns from a matrix
}
\keyword{internal}
