% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_point.R
\name{.wbs_process_intervals}
\alias{.wbs_process_intervals}
\title{Process intervals recursively to detect change points}
\usage{
.wbs_process_intervals(full_dist, type, s, e, M, B, alpha, num_cores = NULL)
}
\arguments{
\item{full_dist}{Full distance matrix}

\item{type}{Type of distance/kernel}

\item{s}{Start index of overall interval}

\item{e}{End index of overall interval}

\item{M}{Number of random intervals to generate}

\item{B}{Number of permutations for p-value calculation}

\item{alpha}{Significance level}

\item{num_cores}{Number of cores for parallel processing}
}
\value{
List with detected change points and p-values
}
\description{
Process intervals recursively to detect change points
}
\keyword{internal}
