% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_point.R
\name{.wbs_test_interval}
\alias{.wbs_test_interval}
\title{Test for a change point within a single interval}
\usage{
.wbs_test_interval(full_dist, type, start, end, M, B, num_cores = 1)
}
\arguments{
\item{full_dist}{Full distance matrix}

\item{type}{Type of distance/kernel}

\item{start}{Start index of overall interval}

\item{end}{End index of overall interval}

\item{M}{Number of random intervals to generate}

\item{B}{Number of permutations for p-value calculation}

\item{num_cores}{Number of cores for parallel processing}
}
\value{
List with detected change point and p-value
}
\description{
Test for a change point within a single interval
}
\keyword{internal}
