% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Joint_dep.R
\name{jhsic}
\alias{jhsic}
\title{Joint Hilbert-Schmidt Independence Criterion}
\usage{
jhsic(
  x,
  cc = 1,
  type = "euclidean",
  stat_type = "V",
  bw = NULL,
  expo = 1,
  scale_factor = 0.5,
  group = NULL
)
}
\arguments{
\item{x}{List of matrices or data frames}

\item{cc}{Constant parameter, default is 1}

\item{type}{Type of kernel or distance}

\item{stat_type}{Type of statistic ("V", "U", "US", or "UR")}

\item{bw}{Bandwidth parameter}

\item{expo}{Exponent parameter}

\item{scale_factor}{Scaling factor for bandwidth}

\item{group}{Optional grouping of variables}
}
\value{
Joint HSIC value
}
\description{
Calculates the joint HSIC for multiple datasets.
}
