% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditional_tests.R
\name{kcid}
\alias{kcid}
\title{Kernel Conditional Independence Test}
\usage{
kcid(
  data_x,
  data_y,
  data_z,
  type_x = "gaussian",
  type_y = "gaussian",
  type_z = "gaussian",
  bw_x = NULL,
  bw_y = NULL,
  bw_z = NULL,
  expo_x = 1,
  expo_y = 1,
  expo_z = 1,
  scale_factor = 0.5,
  group_x = NULL,
  group_y = NULL,
  group_z = NULL,
  method = c("bootstrap", "knn"),
  B = 299,
  knn_k = 5,
  knn_weighted = FALSE,
  num_cores = parallel::detectCores() - 1
)
}
\arguments{
\item{data_x}{First dataset}

\item{data_y}{Second dataset}

\item{data_z}{Conditioning dataset}

\item{type_x}{Type of kernel for first dataset}

\item{type_y}{Type of kernel for second dataset}

\item{type_z}{Type of kernel for conditioning dataset}

\item{bw_x}{Bandwidth parameter for first dataset}

\item{bw_y}{Bandwidth parameter for second dataset}

\item{bw_z}{Bandwidth parameter for conditioning dataset}

\item{expo_x}{Exponent parameter for first dataset}

\item{expo_y}{Exponent parameter for second dataset}

\item{expo_z}{Exponent parameter for conditioning dataset}

\item{scale_factor}{Scaling factor for bandwidth}

\item{group_x}{Grouping for first dataset variables}

\item{group_y}{Grouping for second dataset variables}

\item{group_z}{Grouping for conditioning dataset variables}

\item{method}{Resampling method: "bootstrap" or "knn"}

\item{B}{Number of bootstrap replicates}

\item{knn_k}{Number of nearest neighbors for KNN method}

\item{knn_weighted}{Whether to use weighted KNN}

\item{num_cores}{Number of cores for parallel computing}
}
\value{
A list with p-value, test statistic, and other results
}
\description{
Tests for conditional independence between two variables given a third variable.
}
