% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-kdist.R
\docType{package}
\name{kdist}
\alias{kdist}
\alias{kdist-package}
\title{kdist: A package for calculating and plotting non-Gaussian distributions}
\description{
The kdist package provides two categories of important functions:
dk etc, and weiplot.
}
\section{dk functions}{

The k{kdist} functions dk, pk, qk and rk, calculates the K-distribution
}

\section{weiplot functions}{

weiplot takes data and plots it on Weibull paper. Weilines adds lines
to a Weibull plot.
}

