% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditional_tests.R
\name{knn_conditional_sampling}
\alias{knn_conditional_sampling}
\title{K-Nearest Neighbor Conditional Sampling}
\usage{
knn_conditional_sampling(z_data, k = 5, weighted = TRUE)
}
\arguments{
\item{z_data}{Conditioning dataset}

\item{k}{Number of nearest neighbors}

\item{weighted}{Whether to use weighted sampling}
}
\value{
A vector of sampled indices
}
\description{
Performs KNN-based conditional sampling.
}
