% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Joint_dep.R
\name{mhsic_parallel}
\alias{mhsic_parallel}
\title{A Parallel Implementation of HSIC for Mutual Independence Testing in High-dimension}
\usage{
mhsic_parallel(
  x,
  type = "euclidean",
  bw = NULL,
  expo = 1,
  scale_factor = 0.5,
  cores = NULL
)
}
\arguments{
\item{x}{Matrix with columns to test for mutual independence}

\item{type}{Type of kernel or distance}

\item{bw}{Bandwidth parameter}

\item{expo}{Exponent parameter}

\item{scale_factor}{Scaling factor for bandwidth}

\item{cores}{Number of cores for parallel processing}
}
\value{
A list containing test statistic and p-value
}
\description{
A Parallel Implementation of HSIC for Mutual Independence Testing in High-dimension
}
