% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_point.R
\name{mmd_recur}
\alias{mmd_recur}
\title{Calculate MMD Recursively}
\usage{
mmd_recur(full_dist, type)
}
\arguments{
\item{full_dist}{Full distance matrix}

\item{type}{Type of distance}
}
\value{
A vector of MMD values
}
\description{
Calculates MMD values recursively for various split points.
}
