% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_point.R
\name{my_function}
\alias{my_function}
\title{Retrieve a statistic from the internal asymptotic_stats list}
\usage{
my_function(key)
}
\arguments{
\item{key}{A character string specifying the name of the statistic to retrieve.}
}
\value{
The corresponding statistic (could be a number, list, etc.).
}
\description{
This function retrieves a statistic from the internal dataset \code{asymptotic_stats}
by a given key.
}
\examples{
\dontrun{
my_function("mean_stat")
}

}
