% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot.dcov_test}
\alias{plot.dcov_test}
\title{Plot a histogram of permutation values for a dcov_test}
\usage{
\method{plot}{dcov_test}(x, ...)
}
\arguments{
\item{x}{A dcov_test object}

\item{...}{Additional arguments to be passed to methods}
}
\value{
Invisibly returns the plot object
}
\description{
Plot a histogram of permutation values for a dcov_test
}
