% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot.dhsic_test}
\alias{plot.dhsic_test}
\title{Plot method for dhsic_test}
\usage{
\method{plot}{dhsic_test}(x, ...)
}
\arguments{
\item{x}{A dhsic_test object}

\item{...}{Additional arguments to be passed to methods}
}
\value{
Invisibly returns the plot object
}
\description{
Plot method for dhsic_test
}
