% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot.ed_test}
\alias{plot.ed_test}
\title{Plot a histogram of permutation values for a ed_test}
\usage{
\method{plot}{ed_test}(x, ...)
}
\arguments{
\item{x}{An ed_test object}

\item{...}{Additional arguments to be passed to methods}
}
\value{
Invisibly returns the plot object
}
\description{
Plot a histogram of permutation values for a ed_test
}
