% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot.hsic_test}
\alias{plot.hsic_test}
\title{Plot method for hsic_test}
\usage{
\method{plot}{hsic_test}(x, ...)
}
\arguments{
\item{x}{A hsic_test object}

\item{...}{Additional arguments to be passed to methods}
}
\value{
Invisibly returns the plot object
}
\description{
Plot method for hsic_test
}
