% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot.jhsic_test}
\alias{plot.jhsic_test}
\title{Plot method for jhsic_test}
\usage{
\method{plot}{jhsic_test}(x, ...)
}
\arguments{
\item{x}{A jhsic_test object}

\item{...}{Additional arguments to be passed to methods}
}
\value{
Invisibly returns the plot object
}
\description{
Plot method for jhsic_test
}
