% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot.mdd_test}
\alias{plot.mdd_test}
\title{Plot a histogram of bootstrap values for mdd_test}
\usage{
\method{plot}{mdd_test}(x, ...)
}
\arguments{
\item{x}{A mdd_test object}

\item{...}{Additional arguments to be passed to methods}
}
\value{
Invisibly returns the plot object
}
\description{
Plot a histogram of bootstrap values for mdd_test
}
