% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot.mmd_test}
\alias{plot.mmd_test}
\title{Plot a histogram of permutation values for independence tests}
\usage{
\method{plot}{mmd_test}(x, ...)
}
\arguments{
\item{x}{A test object (mmd_test, hsic_test, dhsic_test, or jhsic_test)}

\item{...}{Additional arguments to be passed to methods}
}
\value{
Invisibly returns the plot object
}
\description{
Plot a histogram of permutation values for independence tests
}
