% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.dcov_test}
\alias{print.dcov_test}
\title{Print a summary of a dcov_test object}
\usage{
\method{print}{dcov_test}(x, ...)
}
\arguments{
\item{x}{A dcov_test object}

\item{...}{Additional arguments to be passed to methods}
}
\value{
Invisibly returns the object
}
\description{
Print a summary of a dcov_test object
}
