% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.dhsic_test}
\alias{print.dhsic_test}
\title{Print a summary of a dhsic_test object}
\usage{
\method{print}{dhsic_test}(x, ...)
}
\arguments{
\item{x}{A dhsic_test object}

\item{...}{Additional arguments to be passed to methods}
}
\value{
Invisibly returns the object
}
\description{
Print a summary of a dhsic_test object
}
