% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.ed_test}
\alias{print.ed_test}
\title{Print a summary of a ed_test object}
\usage{
\method{print}{ed_test}(x, ...)
}
\arguments{
\item{x}{An ed_test object}

\item{...}{Additional arguments to be passed to methods}
}
\value{
Invisibly returns the object
}
\description{
Print a summary of a ed_test object
}
