% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.hsic_test}
\alias{print.hsic_test}
\title{Print a summary of a hsic_test object}
\usage{
\method{print}{hsic_test}(x, ...)
}
\arguments{
\item{x}{A hsic_test object}

\item{...}{Additional arguments to be passed to methods}
}
\value{
Invisibly returns the object
}
\description{
Print a summary of a hsic_test object
}
