% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.jhsic_test}
\alias{print.jhsic_test}
\title{Print a summary of a jhsic_test object}
\usage{
\method{print}{jhsic_test}(x, ...)
}
\arguments{
\item{x}{A jhsic_test object}

\item{...}{Additional arguments to be passed to methods}
}
\value{
Invisibly returns the object
}
\description{
Print a summary of a jhsic_test object
}
