% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.mdd_hd_test}
\alias{print.mdd_hd_test}
\title{Print a summary of a mdd_hd_test object}
\usage{
\method{print}{mdd_hd_test}(x, ...)
}
\arguments{
\item{x}{A mdd_hd_test object}

\item{...}{Additional arguments to be passed to methods}
}
\value{
Invisibly returns the object
}
\description{
Print a summary of a mdd_hd_test object
}
