% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.mmd_test}
\alias{print.mmd_test}
\title{Print a summary of a mmd_test object}
\usage{
\method{print}{mmd_test}(x, ...)
}
\arguments{
\item{x}{A mmd_test object}

\item{...}{Additional arguments to be passed to methods}
}
\value{
Invisibly returns the object
}
\description{
Print a summary of a mmd_test object
}
