% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_point.R
\name{stat_recur}
\alias{stat_recur}
\title{Calculate Statistics Recursively}
\usage{
stat_recur(full_dist, type)
}
\arguments{
\item{full_dist}{Full distance matrix}

\item{type}{Type of distance}
}
\value{
A list with test statistic, change point, and all statistics
}
\description{
Calculates test statistics for change point detection recursively.
This is an internal function used by kcpd_single.
}
