% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{transform_kernel_to_distance}
\alias{transform_kernel_to_distance}
\title{Transform Kernel Matrix to Distance Matrix}
\usage{
transform_kernel_to_distance(K)
}
\arguments{
\item{K}{Kernel matrix}
}
\value{
Distance matrix
}
\description{
Helper function to transform a kernel matrix to a distance matrix.
}
\keyword{internal}
