% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_point.R
\name{adjustedRandIndex}
\alias{adjustedRandIndex}
\title{Adjusted Rand Index}
\usage{
adjustedRandIndex(x, y)
}
\arguments{
\item{x}{First clustering}

\item{y}{Second clustering}
}
\value{
Adjusted Rand Index value
}
\description{
Calculates the Adjusted Rand Index between two clusterings.
}
