% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KDist.R
\name{bw_optim}
\alias{bw_optim}
\title{Calculate optimal bandwidth}
\usage{
bw_optim(x, sample_limit = 1000, scale_factor = 0.5, group = NULL)
}
\arguments{
\item{x}{Matrix or vector of data}

\item{sample_limit}{Maximum number of samples to use for calculation}

\item{scale_factor}{Scaling factor for the bandwidth}

\item{group}{Optional grouping of variables}
}
\value{
Bandwidth parameter value or vector
}
\description{
Calculate optimal bandwidth
}
