% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KDist.R
\name{chsic}
\alias{chsic}
\title{Cross-Sample Hilbert-Schmidt Independence Criterion}
\usage{
chsic(
  x,
  y = NULL,
  type = "euclidean",
  bw = NULL,
  expo = 1,
  scale_factor = 0.5,
  group = NULL
)
}
\arguments{
\item{x}{First sample or cross-distance matrix}

\item{y}{Second sample (optional)}

\item{type}{Type of kernel or distance}

\item{bw}{Bandwidth parameter}

\item{expo}{Exponent parameter}

\item{scale_factor}{Scaling factor for bandwidth}

\item{group}{Optional grouping of variables}
}
\value{
Cross HSIC value
}
\description{
Calculates the cross HSIC between two samples.
}
