% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Joint_dep.R
\name{dhsic}
\alias{dhsic}
\title{d-variable Hilbert-Schmidt Independence Criterion (dHSIC)}
\usage{
dhsic(
  x,
  type = "gaussian",
  bw = NULL,
  expo = 1,
  scale_factor = 0.5,
  group = NULL
)
}
\arguments{
\item{x}{List of matrices or data frames}

\item{type}{Type of kernel or distance}

\item{bw}{Bandwidth parameter}

\item{expo}{Exponent parameter}

\item{scale_factor}{Scaling factor for bandwidth}

\item{group}{Optional grouping of variables}
}
\value{
dHSIC value
}
\description{
Calculate the dHSIC for multiple datasets.
}
