% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_point.R
\name{get_seeded_intervals}
\alias{get_seeded_intervals}
\title{Generate Seeded Intervals}
\usage{
get_seeded_intervals(n, decay = sqrt(2), unique.int = FALSE, bound = 2)
}
\arguments{
\item{n}{Length of time series}

\item{decay}{Decay factor for interval generation}

\item{unique.int}{Whether to return unique intervals}

\item{bound}{Minimum interval size}
}
\value{
A matrix of interval start and end points
}
\description{
Creates seeded intervals for binary segmentation.
}
