% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesis_tests.R
\name{hd_two_test}
\alias{hd_two_test}
\title{High-Dimensional Two-Sample Test}
\usage{
hd_two_test(
  x,
  y,
  type = "e-dist",
  bw = NULL,
  expo = 1,
  scale_factor = 0.5,
  group = NULL
)
}
\arguments{
\item{x}{First sample}

\item{y}{Second sample}

\item{type}{Type of kernel or distance}

\item{bw}{Bandwidth parameter}

\item{expo}{Exponent parameter}

\item{scale_factor}{Scaling factor for bandwidth}

\item{group}{Optional grouping of variables}
}
\value{
A list with test statistic and p-value
}
\description{
Performs a test for the difference between two high-dimensional distributions.
}
