% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KDist.R
\name{hsic}
\alias{hsic}
\title{Hilbert-Schmidt Independence Criterion (HSIC)}
\usage{
hsic(
  x,
  y,
  type = "euclidean",
  bw = NULL,
  expo = 1,
  scale_factor = 0.5,
  group = NULL,
  u_center = FALSE,
  is_distance = FALSE
)
}
\arguments{
\item{x}{First dataset or distance matrix}

\item{y}{Second dataset or distance matrix}

\item{type}{Type of kernel or distance}

\item{bw}{Bandwidth parameter}

\item{expo}{Exponent parameter}

\item{scale_factor}{Scaling factor for bandwidth}

\item{group}{Optional grouping of variables}

\item{u_center}{Logical; use U-centering instead of V-centering}

\item{is_distance}{Logical; whether input matrices are already distance matrices}
}
\value{
HSIC value
}
\description{
Calculates the HSIC between two datasets or matrices.
}
