% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KDist.R
\name{mmd}
\alias{mmd}
\title{Calculate Maximum Mean Discrepancy (MMD)}
\usage{
mmd(
  x,
  y = NULL,
  type = "euclidean",
  bw = NULL,
  expo = 1,
  scale_factor = 0.5,
  group = NULL,
  u_center = FALSE,
  n = NULL,
  m = NULL
)
}
\arguments{
\item{x}{First sample or full distance matrix if y is NULL}

\item{y}{Second sample (optional)}

\item{type}{Type of kernel or distance}

\item{bw}{Bandwidth parameter}

\item{expo}{Exponent parameter}

\item{scale_factor}{Scaling factor for bandwidth}

\item{group}{Optional grouping of variables}

\item{u_center}{Logical; use U-centering instead of V-centering}

\item{n}{Size of first sample (required if x is a distance matrix)}

\item{m}{Size of second sample (required if x is a distance matrix)}
}
\value{
MMD value
}
\description{
Calculate Maximum Mean Discrepancy (MMD)
}
