% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditional_tests.R
\name{tcdt}
\alias{tcdt}
\title{Two-Sample Conditional Distribution Test}
\usage{
tcdt(
  X1,
  X2,
  Y1,
  Y2,
  x0 = NULL,
  B = 299,
  h = NULL,
  adj_bw = 0.1,
  const_factor = 1,
  h_boot = NULL,
  adj_bw_boot = 0,
  const_factor_boot = 1,
  scale_factor = 1,
  kern_smooth = c("gaussian", "epanechnikov", "uniform"),
  nu = 2,
  stat = c("cmmd", "cged"),
  kern_mmd = c("gaussian", "laplacian"),
  sampling_method = c("bootstrap", "knn"),
  knn_k = 5,
  knn_weighted = FALSE,
  num_cores = parallel::detectCores() - 1
)
}
\arguments{
\item{X1}{First conditioning dataset}

\item{X2}{Second conditioning dataset}

\item{Y1}{First response dataset}

\item{Y2}{Second response dataset}

\item{x0}{Local point for local testing (NULL for global testing)}

\item{B}{Number of bootstrap replications}

\item{h}{Bandwidth for local smoothing of X}

\item{adj_bw}{Adjustment factor for automatic bandwidth}

\item{const_factor}{Constant factor for automatic bandwidth}

\item{h_boot}{Bandwidth for bootstrap}

\item{adj_bw_boot}{Adjustment factor for bootstrap bandwidth}

\item{const_factor_boot}{Constant factor for bootstrap bandwidth}

\item{scale_factor}{Scaling factor for bandwidth}

\item{kern_smooth}{Kernel type for smoothing}

\item{nu}{Order of smoothing kernel}

\item{stat}{Test statistic type}

\item{kern_mmd}{Kernel type for MMD}

\item{sampling_method}{Sampling method}

\item{knn_k}{Number of nearest neighbors for KNN}

\item{knn_weighted}{Whether to use weighted KNN}

\item{num_cores}{Number of cores for parallel computing}
}
\value{
A list with test results
}
\description{
Performs a test for differences in conditional distributions.
}
